chore: move unread banner to separate component
